Created game_play file.
